+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/Serializer/ProtobufOtlpTraceSerializer.cs
@@ -164,14 +164,13 @@ internal static class ProtobufOtlpTraceSerializer
                 {
                     if (otlpTagWriterState.TagCount < maxAttributeCount)
                     {
-                        otlpTagWriterState.WritePosition = ProtobufSerializer.WriteTag(otlpTagWriterState.Buffer, otlpTagWriterState.WritePosition, ProtobufOtlpCommonFieldNumberConstants.InstrumentationScope_Attributes, ProtobufWireType.LEN);
-                        var instrumentationScopeAttributesLengthPosition = otlpTagWriterState.WritePosition;
-                        otlpTagWriterState.WritePosition += ReserveSizeForLength;
+                        ProtobufOtlpTagWriter.WriteKeyValue(
+                            ref otlpTagWriterState,
+                            ProtobufOtlpCommonFieldNumberConstants.InstrumentationScope_Attributes,
+                            activitySourceTagsList[i].Key,
+                            activitySourceTagsList[i].Value,
+                            maxAttributeValueLength);
 
-                        ProtobufOtlpTagWriter.Instance.TryWriteTag(ref otlpTagWriterState, activitySourceTagsList[i].Key, activitySourceTagsList[i].Value, maxAttributeValueLength);
-
-                        var instrumentationScopeAttributesLength = otlpTagWriterState.WritePosition - (instrumentationScopeAttributesLengthPosition + ReserveSizeForLength);
-                        ProtobufSerializer.WriteReservedLength(otlpTagWriterState.Buffer, instrumentationScopeAttributesLengthPosition, instrumentationScopeAttributesLength);
                         otlpTagWriterState.TagCount++;
                     }
                     else
@@ -186,14 +185,13 @@ internal static class ProtobufOtlpTraceSerializer
                 {
                     if (otlpTagWriterState.TagCount < maxAttributeCount)
                     {
-                        otlpTagWriterState.WritePosition = ProtobufSerializer.WriteTag(otlpTagWriterState.Buffer, otlpTagWriterState.WritePosition, ProtobufOtlpCommonFieldNumberConstants.InstrumentationScope_Attributes, ProtobufWireType.LEN);
-                        var instrumentationScopeAttributesLengthPosition = otlpTagWriterState.WritePosition;
-                        otlpTagWriterState.WritePosition += ReserveSizeForLength;
+                        ProtobufOtlpTagWriter.WriteKeyValue(
+                            ref otlpTagWriterState,
+                            ProtobufOtlpCommonFieldNumberConstants.InstrumentationScope_Attributes,
+                            tag.Key,
+                            tag.Value,
+                            maxAttributeValueLength);
 
-                        ProtobufOtlpTagWriter.Instance.TryWriteTag(ref otlpTagWriterState, tag.Key, tag.Value, maxAttributeValueLength);
-
-                        var instrumentationScopeAttributesLength = otlpTagWriterState.WritePosition - (instrumentationScopeAttributesLengthPosition + ReserveSizeForLength);
-                        ProtobufSerializer.WriteReservedLength(otlpTagWriterState.Buffer, instrumentationScopeAttributesLengthPosition, instrumentationScopeAttributesLength);
                         otlpTagWriterState.TagCount++;
                     }
                     else
@@ -255,8 +253,9 @@ internal static class ProtobufOtlpTraceSerializer
         writePosition = WriteTraceFlags(buffer, writePosition, activity.ActivityTraceFlags, activity.HasRemoteParent, ProtobufOtlpTraceFieldNumberConstants.Span_Flags);
         writePosition = ProtobufSerializer.WriteStringWithTag(buffer, writePosition, ProtobufOtlpTraceFieldNumberConstants.Span_Name, activity.DisplayName);
         writePosition = ProtobufSerializer.WriteEnumWithTag(buffer, writePosition, ProtobufOtlpTraceFieldNumberConstants.Span_Kind, (int)activity.Kind + 1);
-        writePosition = ProtobufSerializer.WriteFixed64WithTag(buffer, writePosition, ProtobufOtlpTraceFieldNumberConstants.Span_Start_Time_Unix_Nano, (ulong)activity.StartTimeUtc.ToUnixTimeNanoseconds());
-        writePosition = ProtobufSerializer.WriteFixed64WithTag(buffer, writePosition, ProtobufOtlpTraceFieldNumberConstants.Span_End_Time_Unix_Nano, (ulong)(activity.StartTimeUtc.ToUnixTimeNanoseconds() + activity.Duration.ToNanoseconds()));
+        var startTimeUnixNano = activity.StartTimeUtc.ToUnixTimeNanoseconds();
+        writePosition = ProtobufSerializer.WriteFixed64WithTag(buffer, writePosition, ProtobufOtlpTraceFieldNumberConstants.Span_Start_Time_Unix_Nano, (ulong)startTimeUnixNano);
+        writePosition = ProtobufSerializer.WriteFixed64WithTag(buffer, writePosition, ProtobufOtlpTraceFieldNumberConstants.Span_End_Time_Unix_Nano, (ulong)(startTimeUnixNano + activity.Duration.ToNanoseconds()));
 
         (writePosition, var statusCode, var statusMessage) = WriteActivityTags(buffer, writePosition, sdkLimitOptions, activity);
         writePosition = WriteSpanEvents(buffer, writePosition, sdkLimitOptions, activity);
@@ -338,13 +337,13 @@ internal static class ProtobufOtlpTraceSerializer
 
             if (otlpTagWriterState.TagCount < maxAttributeCount)
             {
-                otlpTagWriterState.WritePosition = ProtobufSerializer.WriteTag(otlpTagWriterState.Buffer, otlpTagWriterState.WritePosition, ProtobufOtlpTraceFieldNumberConstants.Span_Attributes, ProtobufWireType.LEN);
-                var spanAttributesLengthPosition = otlpTagWriterState.WritePosition;
-                otlpTagWriterState.WritePosition += ReserveSizeForLength;
+                ProtobufOtlpTagWriter.WriteKeyValue(
+                    ref otlpTagWriterState,
+                    ProtobufOtlpTraceFieldNumberConstants.Span_Attributes,
+                    tag.Key,
+                    tag.Value,
+                    maxAttributeValueLength);
 
-                ProtobufOtlpTagWriter.Instance.TryWriteTag(ref otlpTagWriterState, tag.Key, tag.Value, maxAttributeValueLength);
-
-                ProtobufSerializer.WriteReservedLength(buffer, spanAttributesLengthPosition, otlpTagWriterState.WritePosition - (spanAttributesLengthPosition + 4));
                 otlpTagWriterState.TagCount++;
             }
             else
@@ -414,11 +413,13 @@ internal static class ProtobufOtlpTraceSerializer
         {
             if (otlpTagWriterState.TagCount < maxAttributeCount)
             {
-                otlpTagWriterState.WritePosition = ProtobufSerializer.WriteTag(otlpTagWriterState.Buffer, otlpTagWriterState.WritePosition, ProtobufOtlpTraceFieldNumberConstants.Event_Attributes, ProtobufWireType.LEN);
-                var eventAttributesLengthPosition = otlpTagWriterState.WritePosition;
-                otlpTagWriterState.WritePosition += ReserveSizeForLength;
-                ProtobufOtlpTagWriter.Instance.TryWriteTag(ref otlpTagWriterState, tag.Key, tag.Value, maxAttributeValueLength);
-                ProtobufSerializer.WriteReservedLength(buffer, eventAttributesLengthPosition, otlpTagWriterState.WritePosition - (eventAttributesLengthPosition + ReserveSizeForLength));
+                ProtobufOtlpTagWriter.WriteKeyValue(
+                    ref otlpTagWriterState,
+                    ProtobufOtlpTraceFieldNumberConstants.Event_Attributes,
+                    tag.Key,
+                    tag.Value,
+                    maxAttributeValueLength);
+
                 otlpTagWriterState.TagCount++;
             }
             else
@@ -496,11 +497,13 @@ internal static class ProtobufOtlpTraceSerializer
         {
             if (otlpTagWriterState.TagCount < maxAttributeCount)
             {
-                otlpTagWriterState.WritePosition = ProtobufSerializer.WriteTag(otlpTagWriterState.Buffer, otlpTagWriterState.WritePosition, ProtobufOtlpTraceFieldNumberConstants.Link_Attributes, ProtobufWireType.LEN);
-                var linkAttributesLengthPosition = otlpTagWriterState.WritePosition;
-                otlpTagWriterState.WritePosition += ReserveSizeForLength;
-                ProtobufOtlpTagWriter.Instance.TryWriteTag(ref otlpTagWriterState, tag.Key, tag.Value, maxAttributeValueLength);
-                ProtobufSerializer.WriteReservedLength(buffer, linkAttributesLengthPosition, otlpTagWriterState.WritePosition - (linkAttributesLengthPosition + ReserveSizeForLength));
+                ProtobufOtlpTagWriter.WriteKeyValue(
+                    ref otlpTagWriterState,
+                    ProtobufOtlpTraceFieldNumberConstants.Link_Attributes,
+                    tag.Key,
+                    tag.Value,
+                    maxAttributeValueLength);
+
                 otlpTagWriterState.TagCount++;
             }
             else